apps/proxy/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
fastapi-limiter==0.1.5
redis==5.0.1
web3==6.11.0
cryptography==41.0.7